Fixes required for windows functionality
Windows can't handle TTY, so this checks the OS before setting the flag. Allows the user to configure the timeout, because 60 seconds is annoying if you're setting a lot of variables. Also allows parameters to be passed into the editor call, as Visual Studio Code, for example, requires a -w flag to block the terminal until the user is done editing.

Co-Authored-By: andrewkurzweil <26608741+andrewkurzweil@users.noreply.github.com>